refactor: rename DataType to ColType and add a ColumnType method to semantic.Type
This renames and adds a GoDoc for ColType to make it clear that the
DataType is only for column data types. It also adds a ColumnType method
to the semantic types so that it is possible to take a value and
determine the type that should be used for a column.