refactor: add some new table builder interface ideas
The current implementation of the table builder feels like it is a
one-size fits all approach, but it feels like it requires too much
programmer hand-holding to get the desired behavior which could increase
the proliferation of bugs and cause extra mental load.

This commit creates a new package and defines a set of interfaces for
constructing tables in the two different methods: row or column based.