feat: initial work on a scanner using ragel
The current parser is a mix of the lexing and parsing components and it
is decently slow and doesn't follow the spec completely. This is the
initial process of creating a well tested scanner for flux.

It uses ragel for the finite state machine generation. It supports
lexing the entirety of the language. It includes two methods for
scanning the input because, in the grammar, regular expressions and the
division operator will conflict with each other. The `ScanNoRegex`
method is intended to be used in contexts where we are not expecting a
regular expression (such as in an expression where the next token should
be an operator).

This package is placed into internal so that it is not used and it can
be gradually modified if needed while developing the parser. It may be
moved to a non-internal package in the future.