simple parser implementation
This is an idea for how to lay the groundwork for a parser that will
also be useful for both code completion and better error messages
(hopefully).

This is a problem I ran into frequently when trying to use influxql
outside of the actual database. Interactive use meant that the parser
itself wasn't interruptible. This parser is built in a way where it can
partially consume an input and then return a checkpoint for where it is
currently located. Normally, this would require the stack and it
wouldn't be possible to continue without starting a goroutine and
running the parser within that goroutine.

This also may make it possible to build the repl in a way where it will
automatically execute commands, but if it sees something like `|>` at
the end, it will see the statement is incomplete and that it should
continue waiting for the parser to be sent more input.