refactor a bunch of stuff and split into separate files
the system is a little bit different now. instead of parse returning
false when an error happens, it prints true since it consumed a token
and determined there was an error. when it doesn't consume a token or it
reached EOF, it then says it didn't consume anything.

when an error happens, it will refuse to check for new tokens so it is
up to the caller to either invoke Get to retrieve the object (if they
don't care or can't do anything about EOF) or they can call IsTerminal
to determine if they are in a terminal state.

It is up to the caller to figure that out. The program struct now uses
that functionality though since it needs to process a list.