refactor: modify the FormatOption interface
FormatOption was previously defined as a function with an unexported
parameter which kept it as private. After looking at a few other
libraries that use the same pattern, I found it was better for future
compatibiity if FormatOption was changed into an interface and the
implementation of it being a function was hidden behind that.

This new method is how `zap` implements options and how `go-cmp` also
does it. When implemented in this way, the function that returns a
FormatOption is right next to the implementation so it is easy to find
what the options are without looking at individual function signatures.

It's small, but I think it makes the publicly exposed API more friendly.