refactor: define the array interface further
This specifies the array interface further to specify that the builder
will create an immutable reference and attempts to define what happens
with the reference the builder keeps.

This was an aspect of the interface that I hadn't really specified
because I wasn't sure what would be needed, but the basics are that the
builder will either copy the data or keep a reference to the data when
constructing an array. If it is modified afterwards, it will either
modify its own internal buffer or perform a copy-on-write technique.

Both of these implementations have the same outward appearance so the
docs for the methods focus on that rather than the specific
implementation.